Le kill marche avec gregaire

+++ KillHandler.cpp
@@ -11,28 +11,35 @@ KillHandler::~KillHandler(){
     
 }
 
-bool KillHandler::kill(Bestiole& bestiole, Milieu * monMilieu){
-    if (bestiole.getAge() >= bestiole.getAgeLimite()) {
+bool KillHandler::kill(Bestiole* bestiole, Milieu * monMilieu){
+    if (bestiole->getAge() >= bestiole->getAgeLimite()) {
+        std::cout << "Kill dû à l'age. Elle avait " << bestiole->getAge() << "ans et son age limite est " << bestiole->getAgeLimite() << std::endl;
         monMilieu-> removeMember(bestiole);
+        return true;
     }
 
     //Check for collision with other bestioles
-    Bestiole collidedBestiole = monMilieu->checkCollision(bestiole);
-    // TODO : si pas collision -> checkCollision ne renvoie rien -> donc créer une bestiole "NULL" (NULL ne marche pas on l'a déjà testé)
-    //        Ensuite vérifier si colidiedBestiol n'est pas la bestiole "NULL" 
-    //        Sinon détruire colidiedBestiole et ne pas rentrer dans la boucle if rand else
-    // TODO : corriger méthode checkCollision dans Milieu (ligne 110) pour renvoyer une bestioleNULL si non collision
-    // if (collidedBestiole != BestioleNULL) {
+    Bestiole* collidedBestiole = monMilieu->checkCollision(bestiole);
+    
+    //Print bestiole and collidedBestiole positions
+    std::cout << "Bestiole " << bestiole->getIdentite() << " is at (" << bestiole->getX() << "," << bestiole->getY() << ")."<< std::endl;
+    //std::cout << "Collided bestiole " << collidedBestiole->getIdentite() << " is at (" << collidedBestiole->getX() << "," << collidedBestiole->getY() << ")."<< std::endl;
+    
+    if (collidedBestiole != nullptr) {
+        cout << "We get in the collided != nullptr" << endl;
         if(rand()> 0.75){ // We kill the bestioles
             monMilieu -> removeMember(bestiole);
             monMilieu -> removeMember(collidedBestiole);
+            return true;
         }
     
         else{ // Otherwise we just move them away
-            bestiole.setOrientation(-bestiole.getOrientation());
-            collidedBestiole.setOrientation(-collidedBestiole.getOrientation());
+            bestiole->setOrientation(-bestiole->getOrientation());
+            collidedBestiole->setOrientation(-collidedBestiole->getOrientation());
+            return false;
         }
-    //}
+    }
+    return false;
     //else {
     //    delete collidedBestiole;
     //}
